fixedadmin

+++ rate-My-Hike-BackEnd/src/main/java/com/example/ratemyhike/Model/Admin.java
@@ -21,7 +21,7 @@ public class Admin {
         this.email = email;
     }
 
-    public int getAccountNumber() {
+    public long getAccountNumber() {
         return accountNumber;
     }
